refactor: field value to own component
